YQL-19747 Refactor NameSet
